utility/requirements.txt
requests>=2.30.0
python-dotenv>=1.1.0
schedule>=1.2.0